Conflicts:
	cocos/2d/CCLabel.cpp

+++ samples/Cpp/TestCpp/Classes/ShaderTest/ShaderTest.cpp
@@ -547,8 +547,8 @@ void SpriteBlur::onDraw()
     //
     // Attributes
     //
-#define kQuadSize sizeof(_quad.bl)
-    long offset = (long)&_quad;
+    #define kQuadSize sizeof(_quad.bl)
+    size_t offset = (size_t)&_quad;
     
     // vertex
     int diff = offsetof( V3F_C4B_T2F, vertices);
